x11: move scaling determination funcs into xcb_connection.rs
